pyproject.toml
[project]
name = "py-agui"
version = "0.1.0"
description = "Python Agentic UI - Real-time agentic chat interfaces with FastHTML, featuring 3-pane layout with settings, streaming chat, and thinking trace"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Christoffer Bjorkskog", email = "christoffer.bjorkskog@novia.fi" },
    { name = "Tarvo Kaljuvee" }
]
keywords = ["fasthtml", "agui", "pydantic-ai", "chat", "websocket", "agentic-ui", "streaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: User Interfaces",
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-ai>=1.34.0",
    "python-fasthtml>=0.12.36",
    "ag-ui-protocol>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/kaljuvee/py-agui"
Repository = "https://github.com/kaljuvee/py-agui"
Documentation = "https://github.com/kaljuvee/py-agui#readme"
"Bug Reports" = "https://github.com/kaljuvee/py-agui/issues"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["py_agui"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
